test_ncbistr (s_TruncateSpaces): Compensate for use of char32_t.
When reporting on iswspace's limited handling of wide characters, cast
the character to be tested to int to avoid warnings from strict C++
standard library implementations (such as recent GNU versions) that
explicitly block attempts to present them in any fashion via
narrow-character streams.

git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@103175 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/corelib/test/test_ncbistr.cpp
@@ -4467,7 +4467,8 @@ BOOST_AUTO_TEST_CASE(s_TruncateSpaces)
     cout << "Testing iswspace" << endl;
     for (i=0; ws[i]; ++i) {
         if (iswspace(ws[i]) == 0) {
-            cout << "WARNING: wide char " << hex << ws[i] << " is not whitespace" << endl;
+            cout << "WARNING: wide char " << hex << static_cast<int>(ws[i])
+                 << " is not whitespace" << endl;
         }
     }
     cout << "Testing CStringUTF8::IsWhiteSpace" << endl;
